added tests for Admin.php and ProductFeed.php
